+++ pkg/gatewayserver/gatewayserver.go
@@ -45,6 +45,7 @@ import (
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/io/udp"
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/upstream"
 	"go.thethings.network/lorawan-stack/pkg/gatewayserver/upstream/ns"
+	"go.thethings.network/lorawan-stack/pkg/gatewayserver/upstream/packetbroker"
 	"go.thethings.network/lorawan-stack/pkg/log"
 	"go.thethings.network/lorawan-stack/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/pkg/rpcmiddleware/hooks"
@@ -150,6 +151,8 @@ func New(c *component.Component, conf *Config, opts ...Option) (gs *GatewayServe
 		switch name {
 		case "cluster":
 			handler = ns.NewHandler(gs.Context(), c, prefix)
+		case "packetbroker":
+			handler = packetbroker.NewHandler(gs.Context(), c, prefix)
 		default:
 			return nil, errInvalidUpstreamName.WithAttributes("name", name)
 		}