+++ pkg/networkserver/downlink.go
@@ -33,6 +33,7 @@ import (
 	"go.thethings.network/lorawan-stack/pkg/log"
 	"go.thethings.network/lorawan-stack/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/pkg/unique"
+	"google.golang.org/grpc"
 )
 
 // DownlinkTaskQueue represents an entity, that holds downlink tasks sorted by timestamp.
@@ -573,7 +574,7 @@ func (ns *NetworkServer) generateDataDownlink(ctx context.Context, dev *ttnpb.En
 }
 
 type downlinkPath struct {
-	ttnpb.GatewayIdentifiers
+	*ttnpb.GatewayIdentifiers
 	*ttnpb.DownlinkPath
 }
 
@@ -584,29 +585,34 @@ func downlinkPathsFromMetadata(mds ...*ttnpb.RxMetadata) []downlinkPath {
 		return mds[i].SNR > mds[j].SNR
 	})
 	head := make([]downlinkPath, 0, len(mds))
+	body := make([]downlinkPath, 0, len(mds))
 	tail := make([]downlinkPath, 0, len(mds))
 	for _, md := range mds {
 		if len(md.UplinkToken) == 0 || md.DownlinkPathConstraint == ttnpb.DOWNLINK_PATH_CONSTRAINT_NEVER {
 			continue
 		}
-
 		path := downlinkPath{
-			GatewayIdentifiers: md.GatewayIdentifiers,
 			DownlinkPath: &ttnpb.DownlinkPath{
 				Path: &ttnpb.DownlinkPath_UplinkToken{
 					UplinkToken: md.UplinkToken,
 				},
 			},
 		}
-		switch md.DownlinkPathConstraint {
-		case ttnpb.DOWNLINK_PATH_CONSTRAINT_NONE:
-			head = append(head, path)
-
-		case ttnpb.DOWNLINK_PATH_CONSTRAINT_PREFER_OTHER:
+		if md.PacketBroker != nil {
 			tail = append(tail, path)
+		} else {
+			path.GatewayIdentifiers = &md.GatewayIdentifiers
+			switch md.DownlinkPathConstraint {
+			case ttnpb.DOWNLINK_PATH_CONSTRAINT_NONE:
+				head = append(head, path)
+			case ttnpb.DOWNLINK_PATH_CONSTRAINT_PREFER_OTHER:
+				body = append(body, path)
+			}
 		}
 	}
-	return append(head, tail...)
+	res := append(head, body...)
+	res = append(res, tail...)
+	return res
 }
 
 // classAWindowsAvailableAt returns whether class A downlink can be made following up
@@ -713,6 +719,56 @@ func newDataDownlinkScheduleRequest(req *ttnpb.TxRequest, ids ttnpb.EndDeviceIde
 	}
 }
 
+type downlinkTarget interface {
+	Equal(downlinkTarget) bool
+	Schedule(context.Context, *ttnpb.DownlinkMessage, ...grpc.CallOption) (time.Duration, error)
+}
+
+type gatewayServerDownlinkTarget struct {
+	peer cluster.Peer
+}
+
+func (t *gatewayServerDownlinkTarget) Equal(target downlinkTarget) bool {
+	other, ok := target.(*gatewayServerDownlinkTarget)
+	if !ok {
+		return false
+	}
+	return other.peer == t.peer
+}
+
+func (t *gatewayServerDownlinkTarget) Schedule(ctx context.Context, msg *ttnpb.DownlinkMessage, callOpts ...grpc.CallOption) (time.Duration, error) {
+	conn, err := t.peer.Conn()
+	if err != nil {
+		return 0, err
+	}
+	res, err := ttnpb.NewNsGsClient(conn).ScheduleDownlink(ctx, msg, callOpts...)
+	if err != nil {
+		return 0, err
+	}
+	return res.Delay, nil
+}
+
+type packetBrokerDownlinkTarget struct {
+	peer cluster.Peer
+}
+
+func (t *packetBrokerDownlinkTarget) Equal(target downlinkTarget) bool {
+	_, ok := target.(*packetBrokerDownlinkTarget)
+	return ok
+}
+
+func (t *packetBrokerDownlinkTarget) Schedule(ctx context.Context, msg *ttnpb.DownlinkMessage, callOpts ...grpc.CallOption) (time.Duration, error) {
+	conn, err := t.peer.Conn()
+	if err != nil {
+		return 0, err
+	}
+	_, err = ttnpb.NewNsPbaClient(conn).PublishDownlink(ctx, msg, callOpts...)
+	if err != nil {
+		return 0, err
+	}
+	return peeringScheduleDelay, nil
+}
+
 // scheduleDownlinkByPaths attempts to schedule payload b using parameters in req using paths.
 // scheduleDownlinkByPaths discards req.TxRequest.DownlinkPaths and mutates it arbitrarily.
 // scheduleDownlinkByPaths returns the scheduled downlink or error.
@@ -724,31 +780,44 @@ func (ns *NetworkServer) scheduleDownlinkByPaths(ctx context.Context, req *sched
 	logger := log.FromContext(ctx)
 
 	type attempt struct {
-		peer  cluster.Peer
+		downlinkTarget
 		paths []*ttnpb.DownlinkPath
-	}
-	attempts := make([]*attempt, 0, len(paths))
-	lastAttempt := func() *attempt {
-		return attempts[len(attempts)-1]
 	}
 
 	queuedEvents := make([]events.Event, 0, len(paths))
+	attempts := make([]*attempt, 0, len(paths))
 	for _, path := range paths {
-		logger := logger.WithField(
-			"gateway_uid", unique.ID(ctx, path.GatewayIdentifiers),
-		)
-		p, err := ns.GetPeer(ctx, ttnpb.ClusterRole_GATEWAY_SERVER, path.GatewayIdentifiers)
-		if err != nil {
-			logger.WithError(err).Debug("Could not get Gateway Server")
-			continue
+		var target downlinkTarget
+		if path.GatewayIdentifiers != nil {
+			logger := logger.WithFields(log.Fields(
+				"target", "gateway_server",
+				"gateway_uid", unique.ID(ctx, path.GatewayIdentifiers),
+			))
+			peer, err := ns.GetPeer(ctx, ttnpb.ClusterRole_GATEWAY_SERVER, *path.GatewayIdentifiers)
+			if err != nil {
+				logger.WithError(err).Warn("Failed to get Gateway Server peer")
+				continue
+			}
+			target = &gatewayServerDownlinkTarget{peer: peer}
+		} else {
+			logger := logger.WithField("target", "packet_broker_agent")
+			peer, err := ns.GetPeer(ctx, ttnpb.ClusterRole_PACKET_BROKER_AGENT, nil)
+			if err != nil {
+				logger.WithError(err).Warn("Failed to get Packet Broker Agent peer")
+				continue
+			}
+			target = &packetBrokerDownlinkTarget{peer: peer}
 		}
 
 		var a *attempt
-		if len(attempts) > 0 && lastAttempt().peer == p {
-			a = lastAttempt()
-		} else {
+		if len(attempts) > 0 {
+			if last := attempts[len(attempts)-1]; last.Equal(target) {
+				a = last
+			}
+		}
+		if a == nil {
 			a = &attempt{
-				peer: p,
+				downlinkTarget: target,
 			}
 			attempts = append(attempts, a)
 		}
@@ -767,25 +836,25 @@ func (ns *NetworkServer) scheduleDownlinkByPaths(ctx context.Context, req *sched
 			},
 		}
 		queuedEvents = append(queuedEvents, req.AttemptEvent(ctx, req.EndDeviceIdentifiers, down))
-
 		logger.WithField("path_count", len(req.DownlinkPaths)).Debug("Schedule downlink")
-		cc, err := a.peer.Conn()
+		delay, err := a.Schedule(ctx, down, ns.WithClusterAuth())
 		if err != nil {
 			queuedEvents = append(queuedEvents, req.FailEvent(ctx, req.EndDeviceIdentifiers, err))
 			errs = append(errs, err)
 			continue
 		}
-		res, err := ttnpb.NewNsGsClient(cc).ScheduleDownlink(ctx, down, ns.WithClusterAuth())
-		if err != nil {
-			queuedEvents = append(queuedEvents, req.FailEvent(ctx, req.EndDeviceIdentifiers, err))
-			errs = append(errs, err)
-			continue
-		}
-		logger.WithField("transmit_delay", res.Delay).Debug("Scheduled downlink")
+		transmitAt := timeNow().Add(delay)
+		logger.WithFields(log.Fields(
+			"transmission_delay", delay,
+			"transmit_at", transmitAt,
+		)).Debug("Scheduled downlink")
+		queuedEvents = append(queuedEvents, req.SuccessEvent(ctx, req.EndDeviceIdentifiers, &ttnpb.ScheduleDownlinkResponse{
+			Delay: delay,
+		}))
 		return &scheduledDownlink{
 			Message:    down,
-			TransmitAt: timeNow().Add(res.Delay),
-		}, append(queuedEvents, req.SuccessEvent(ctx, req.EndDeviceIdentifiers, res)), nil
+			TransmitAt: transmitAt,
+		}, queuedEvents, nil
 	}
 	return nil, queuedEvents, downlinkSchedulingError(errs)
 }
@@ -1415,7 +1484,7 @@ func (ns *NetworkServer) processDownlinkTask(ctx context.Context) error {
 					paths = make([]downlinkPath, 0, len(fixedPaths))
 					for i := range fixedPaths {
 						paths = append(paths, downlinkPath{
-							GatewayIdentifiers: fixedPaths[i].GatewayIdentifiers,
+							GatewayIdentifiers: &fixedPaths[i].GatewayIdentifiers,
 							DownlinkPath: &ttnpb.DownlinkPath{
 								Path: &ttnpb.DownlinkPath_Fixed{
 									Fixed: &fixedPaths[i],